splits project reconciliation into dedicated routes
  which allows sectional content to manage its own state
  fixes issue #96
  adds BDD coverage
