+++ web/src/app/projects/[id]/discord/page.tsx
@@ -1,0 +1,5 @@
+import ProjectRouteClient from "../ProjectRouteClient";
+
+export default function ProjectDiscordPage() {
+  return <ProjectRouteClient section="discord" />;
+}
