+++ web/src/app/projects/[id]/docs/page.tsx
@@ -1,0 +1,5 @@
+import ProjectRouteClient from "../ProjectRouteClient";
+
+export default function ProjectDocsPage() {
+  return <ProjectRouteClient section="docs" />;
+}
