+++ web/src/app/projects/[id]/dot-project/page.tsx
@@ -1,0 +1,5 @@
+import ProjectRouteClient from "../ProjectRouteClient";
+
+export default function ProjectDotProjectPage() {
+  return <ProjectRouteClient section="dot-project" />;
+}
