+++ web/src/app/projects/[id]/fossa/page.tsx
@@ -1,0 +1,5 @@
+import ProjectRouteClient from "../ProjectRouteClient";
+
+export default function ProjectFossaPage() {
+  return <ProjectRouteClient section="license-checker" />;
+}
