+++ web/src/app/projects/[id]/github/page.tsx
@@ -1,0 +1,5 @@
+import ProjectRouteClient from "../ProjectRouteClient";
+
+export default function ProjectGithubPage() {
+  return <ProjectRouteClient section="legacy" />;
+}
