+++ web/src/app/projects/[id]/mailing-maintainers/page.tsx
@@ -1,0 +1,5 @@
+import ProjectRouteClient from "../ProjectRouteClient";
+
+export default function ProjectMailingMaintainersPage() {
+  return <ProjectRouteClient section="mailing-maintainers" />;
+}
