+++ web/src/app/projects/[id]/mailing-security/page.tsx
@@ -1,0 +1,5 @@
+import ProjectRouteClient from "../ProjectRouteClient";
+
+export default function ProjectMailingSecurityPage() {
+  return <ProjectRouteClient section="mailing-security" />;
+}
