+++ web/src/app/projects/[id]/slack/page.tsx
@@ -1,0 +1,5 @@
+import ProjectRouteClient from "../ProjectRouteClient";
+
+export default function ProjectSlackPage() {
+  return <ProjectRouteClient section="slack" />;
+}
